requirements.txt
asgiref==3.5.2
azure-cognitiveservices-vision-face==0.6.0
azure-common==1.1.28
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
decorator==4.4.2
defusedxml==0.7.1
dj-database-url==0.5.0
Django==4.0.4
django-allauth==0.50.0
django-environ==0.8.1
django-heroku==0.3.1
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
isodate==0.6.1
msrest==0.6.21
numpy==1.22.4
oauthlib==3.2.0
Pillow==9.1.1
proglog==0.1.10
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.4.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tqdm==4.64.0
urllib3==1.26.9
whitenoise==6.1.0
gunicorn